Odoo connector updates: - odoo->frepple: 2 modes now, one to use during online planning, another one for bigger slow-changing data element - frepple->odoo: ability to export incrementally - removing support for connector in odoo v7 - updates for odoo v8 and frepple 3.2

+++ contrib/odoo/addons_v8/frepple/controllers/inbound.py
@@ -19,6 +19,7 @@ import openerp
 import logging
 from datetime import datetime, timedelta
 from xml.etree.cElementTree import iterparse
+from werkzeug.formparser import parse_form_data
 
 logger = logging.getLogger(__name__)
 
@@ -26,8 +27,7 @@ logger = logging.getLogger(__name__)
 class importer(object):
 
   def __init__(self, req, **kwargs):
-    self.req = req
-    self.database = kwargs.get('database', None)
+    self.database = req.httprequest.form.get('database', None)
     authmeth, auth = req.httprequest.headers['authorization'].split(' ', 1)
     if authmeth.lower() != 'basic':
       raise Exception("No authentication header")
@@ -35,18 +35,27 @@ class importer(object):
     user, password = auth.split(':', 1)
     if not self.database or not user or not password:
       raise Exception("Authentication error")
-    if not self.req.session.authenticate(self.database, user, password):
+    if not req.session.authenticate(self.database, user, password):
       raise Exception("Odoo authentication failed")
     if 'language' in kwargs:
       # If not set we use the default language of the user
-      self.req.session.context['lang'] = kwargs['language']
-    self.company = kwargs.get('company', None)
-    self.datafile = kwargs.get('frePPLe plan')
-
+      req.session.context['lang'] = kwargs['language']
+    self.req = req    
+    self.company = req.httprequest.form.get('company', None)
+    self.datafile = req.httprequest.files.get('frePPLe plan')
+    
+    # The mode argument defines different types of runs:
+    #  - Mode 1:
+    #    Export of the complete plan. This first erase all previous frePPLe
+    #    proposals in draft state.
+    #  - Mode 2:
+    #    Incremental export of some proposed transactions from frePPLe.
+    #    In this mode mode we are not erasing any previous proposals.
+    self.mode = req.httprequest.form.get('mode', 1)
+    
 
   def run(self):
     msg = []
-
     # Look up the company id
     company_id = None
     m = self.req.session.model('res.company')
@@ -55,38 +64,39 @@ class importer(object):
     if not company_id:
       raise Exception("Invalid company name argument")
 
-    # Cancel previous draft purchase quotations
-    m = self.req.session.model('purchase.order')
-    ids = m.search(
-      [('state', '=', 'draft'), ('origin', '=', 'frePPLe')],
-      context=self.req.session.context
-      )
-    m.unlink(ids, self.req.session.context)
-    msg.append("Removed %s old draft purchase quotations" % len(ids))
+    if self.mode == 1:
+      # Cancel previous draft purchase quotations
+      m = self.req.session.model('purchase.order')
+      ids = m.search(
+        [('state', '=', 'draft'), ('origin', '=', 'frePPLe')],
+        context=self.req.session.context
+        )
+      m.unlink(ids, self.req.session.context)
+      msg.append("Removed %s old draft purchase quotations" % len(ids))
 
-    # Cancel previous draft procurement orders
-    proc_order = self.req.session.model('procurement.order')
-    ids = proc_order.search(
-      ['|', ('state', '=', 'draft'), ('state', '=', 'cancel'), ('origin', '=', 'frePPLe')],
-      context=self.req.session.context
-      )
-    proc_order.unlink(ids, self.req.session.context)
-    msg.append("Removed %s old draft procurement orders" % len(ids))
+      # Cancel previous draft procurement orders
+      proc_order = self.req.session.model('procurement.order')
+      ids = proc_order.search(
+        ['|', ('state', '=', 'draft'), ('state', '=', 'cancel'), ('origin', '=', 'frePPLe')],
+        context=self.req.session.context
+        )
+      proc_order.unlink(ids, self.req.session.context)
+      msg.append("Removed %s old draft procurement orders" % len(ids))
 
-    # Cancel previous draft manufacturing orders
-    mfg_order = self.req.session.model('mrp.production')
-    ids = mfg_order.search(
-      ['|', ('state', '=', 'draft'), ('state', '=', 'cancel'), ('origin', '=', 'frePPLe')],
-      context=self.req.session.context
-      )
-    mfg_order.unlink(ids, self.req.session.context)
-    msg.append("Removed %s old draft manufacturing orders" % len(ids))
+      # Cancel previous draft manufacturing orders
+      mfg_order = self.req.session.model('mrp.production')
+      ids = mfg_order.search(
+        ['|', ('state', '=', 'draft'), ('state', '=', 'cancel'), ('origin', '=', 'frePPLe')],
+        context=self.req.session.context
+        )
+      mfg_order.unlink(ids, self.req.session.context)
+      msg.append("Removed %s old draft manufacturing orders" % len(ids))
 
     # Parsing the XML data file
     countproc = 0
     countmfg = 0
     for event, elem in iterparse(self.datafile, events=('start', 'end')):
-      if event == 'end' and elem.tag == 'operationplan':
+      if event == 'end' and elem.tag == 'operationplan':        
         uom_id, item_id = elem.get('item').split(',')
         n = elem.get('operation')
         try:
@@ -107,6 +117,8 @@ class importer(object):
             # proc_order.action_confirm([x], context=self.req.session.context) # it is confirmed by default
             # proc_order.action_po_assign([x], context=self.req.session.context) # TODO no idea of what this is yet, other than not available :)
             countproc += 1
+          # TODO Create a distribution order
+          # elif ????:
           else:
             # Create manufacturing order
             x = mfg_order.create({
